Sally  10cd8b0

+++ src/common/dib.cpp
@@ -1,6 +1,6 @@
 ﻿// SPDX-FileCopyrightText: 2023 Open Salamander Authors
+// SPDX-FileCopyrightText: 2026 Sally Authors
 // SPDX-License-Identifier: GPL-2.0-or-later
-// CommentsTranslationProject: TRANSLATED
 
 #include "precomp.h"
 
@@ -8,7 +8,7 @@
 #include <crtdbg.h>
 #include <ostream>
 #include <limits.h>
-#include <commctrl.h> // Needed for LPCOLORMAP.
+#include <commctrl.h> // need LPCOLORMAP
 
 #if defined(_DEBUG) && defined(_MSC_VER) // without passing file+line to 'new' operator, list of memory leaks shows only 'crtdbg.h(552)'
 #define new new (_NORMAL_BLOCK, __FILE__, __LINE__)
@@ -22,8 +22,8 @@
 
 #include "dib.h"
 
-// Workaround for a runtime check failure in the debug build: the original version of the macro casts rgb to WORD,
-// so the runtime check reports data loss in the RED component.
+// protection against runtime check failure in debug version: original macro version casts rgb to WORD,
+// which reports data loss (RED component)
 #undef GetGValue
 #define GetGValue(rgb) ((BYTE)(((rgb) >> 8) & 0xFF))
 
@@ -76,14 +76,14 @@ HBITMAP LoadBitmapAndMapColors(HINSTANCE hInst, HRSRC hRsrc, int mapCount,
     HDC hDCScreen = HANDLES(GetDC(NULL));
     LPBYTE lpBits;
     lpBits = (LPBYTE)(lpBitmap + 1);
-    lpBits += (1 << (lpBitmapInfo->biBitCount)) * sizeof(RGBQUAD);
+    lpBits += ((size_t)1 << lpBitmapInfo->biBitCount) * sizeof(RGBQUAD);
     HBITMAP hbm = HANDLES(CreateDIBitmap(hDCScreen, lpBitmapInfo, CBM_INIT, lpBits,
                                          (LPBITMAPINFO)lpBitmapInfo, DIB_RGB_COLORS));
     if (hbm == NULL)
         TRACE_E("Unable to create bitmap.");
     HANDLES(ReleaseDC(NULL, hDCScreen));
 
-    // free the copy of the bitmap info struct and the resource itself
+    // free copy of bitmap info struct and resource itself
     free(lpBitmapInfo);
     FreeResource(hglb);
 
@@ -245,7 +245,7 @@ DWORD DIBWidth(LPSTR lpDIB)
 //
 // Purpose:    Given a handle to global memory with a DIB spec in it,
 //             and a palette, returns a device dependent bitmap.  The
-//             DDB will be rendered with the specified palette.
+//             The DDB will be rendered with the specified palette.
 //
 // Parms:      hDIB == HANDLE to global memory containing a DIB spec
 //                     (either BITMAPINFOHEADER or BITMAPCOREHEADER)
@@ -308,14 +308,13 @@ HBITMAP DIBToBitmap(HANDLE hDIB, HPALETTE hPal)
     return hBitmap;
 }
 
-// ---------------------------------------------------------------------
+//---------------------------------------------------------------------
 //
 // Function:   MapColor(RGB fromColor, RGB toColor)
 //
-//             Remaps every occurrence of fromColor in the DIB to
-//             toColor.
+//             maps all fromColor colors to toColor color
 //
-// ---------------------------------------------------------------------
+//---------------------------------------------------------------------
 
 int MapColor(HANDLE hDIB, COLORREF fromColor, COLORREF toColor)
 {
