+++ src/plugins/diskmap/DiskMap/Utils.CPixMap.h
@@ -1,4 +1,5 @@
 ﻿// SPDX-FileCopyrightText: 2023 Open Salamander Authors
+// SPDX-FileCopyrightText: 2026 Sally Authors
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
@@ -15,7 +16,7 @@ protected:
 
     void PreparePixMap(int width, int height, BOOL clear)
     {
-        size_t size = width * height * 4;
+        size_t size = (size_t)width * height * 4;
         if (size > this->_pixsize)
         {
             if (this->_pixmap)
